offset for left/right for trapezoidal filter working. doing arithmetic before casting to c_ulong did the trick. also made peaks->Left/Right into its own function, we may want to change it or make it more flexible later on. tested consecutive peaks with decent results. slight baseline shift would persist throughout a lot of the dataset in this case.
2D pixel plotting is working for both a sequence of images (like a crappy video) or just plotting specific time points individually. can clearly see signals e.g. circular blobs moving on the screen.

next, we want to make peak detection and fitting work properly for the whole sensor. some issues that effect this are: really noisy or bad channels, weird pulses that don't work with exponential fit( have a spike followed by a kind of symmetric hump), fitting procedure does not appear robust for all the flavors of pulses we see yet. sometimes the fitted tau seem to round to clean specific numbers, or clearly enough points are not being used to fit certain pulses (3x tau is supposed to be a reasonable fit length). also, consecutive tau with long pulses are a problem.